requirements.txt
requests
tqdm
pandas
psycopg2-binary